+++ scripts/rebuild_specs.js
@@ -85,9 +85,10 @@ function rebuildConverterTests(dirs) {
                 );
                 const serialized = app.serializer.toObject(result);
 
-                const data = JSON.stringify(serialized, null, "  ")
-                    .split(TypeDoc.normalizePath(base))
-                    .join("%BASE%");
+                const data =
+                    JSON.stringify(serialized, null, "  ")
+                        .split(TypeDoc.normalizePath(base))
+                        .join("%BASE%") + "\n";
                 after();
                 fs.writeFileSync(out.replace("dist", "src"), data);
             }