+++ src/lib/output/themes/default/partials/member.declaration.tsx
@@ -1,4 +1,4 @@
-import { DeclarationReflection, ReflectionType } from "../../../../models";
+import { DeclarationReflection, ReflectionKind, ReflectionType } from "../../../../models";
 import { JSX } from "../../../../utils";
 import { renderTypeParametersSignature, wbr } from "../../lib";
 import type { DefaultThemeRenderContext } from "../DefaultThemeRenderContext";
@@ -14,7 +14,7 @@ export const memberDeclaration = (context: DefaultThemeRenderContext, props: Dec
                     {context.type(props.type)}
                 </>
             )}
-            {!!props.defaultValue && (
+            {!!props.defaultValue && props.kind !== ReflectionKind.EnumMember && (
                 <>
                     <span class="tsd-signature-symbol">
                         {" = "}
@@ -30,7 +30,7 @@ export const memberDeclaration = (context: DefaultThemeRenderContext, props: Dec
 
         {!!props.typeParameters && (
             <>
-                <h4 class="tsd-type-parameters-title">Type parameters</h4>
+                <h4 class="tsd-type-parameters-title">Type Parameters</h4>
                 {context.typeParameters(props.typeParameters)}
             </>
         )}