+++ src/lib/output/themes/lib.tsx
@@ -88,9 +88,12 @@ export function renderTypeParametersSignature(
                 <>
                     <span class="tsd-signature-symbol">{"<"}</span>
                     {join(<span class="tsd-signature-symbol">{", "}</span>, typeParameters, (item) => (
-                        <span class="tsd-signature-type" data-tsd-kind={item.kindString}>
-                            {item.name}
-                        </span>
+                        <>
+                            {item.varianceModifier ? `${item.varianceModifier} ` : ""}
+                            <span class="tsd-signature-type" data-tsd-kind={item.kindString}>
+                                {item.name}
+                            </span>
+                        </>
                     ))}
                     <span class="tsd-signature-symbol">{">"}</span>
                 </>